+++ src/util-random.c
@@ -25,9 +25,8 @@
  */
 
 #include "suricata-common.h"
-#include "suricata.h"
 #include "util-random.h"
-#include "util-debug.h"
+
 
 #if !(defined(HAVE_WINCRYPT_H) &&  defined(OS_WIN32))
 #if defined(HAVE_CLOCK_GETTIME)
@@ -38,7 +37,7 @@ static long int RandomGetClock(void)
     clock_gettime(CLOCK_REALTIME, &ts);
 
     // coverity[dont_call : FALSE]
-    srandom((unsigned int)(ts.tv_nsec ^ ts.tv_sec));
+    srandom(ts.tv_nsec ^ ts.tv_sec);
     long int value = random();
     return value;
 }
@@ -103,7 +102,7 @@ long int RandomGet(void)
         return 0;
 
     long int value = 0;
-    ssize_t ret = getrandom(&value, sizeof(value), 0);
+    int ret = getrandom(&value, sizeof(value), 0);
     /* ret should be sizeof(value), but if it is > 0 and < sizeof(value)
      * it's still better than nothing so we return what we have */
     if (ret <= 0) {
